refactor(editor): Extract large chunks from NodeSettings (#17665)
